+++ bin/check-stale-results.rb
@@ -42,11 +42,12 @@ class CheckStaleResults < Sensu::Plugin::Check::CLI
          proc: proc(&:to_i),
          default: nil
 
-  option :ssl,
-         description: 'Use SSL',
-         long: '--ssl',
-         boolean: true,
-         default: false
+  option :protocol,
+         description: 'Protocol for communication http/https',
+         short: '-p PROTOCOL',
+         long: '--protocol PROTOCOL',
+         default: 'http',
+         in: %w(http https)
 
   def initialize
     super
@@ -72,7 +73,7 @@ class CheckStaleResults < Sensu::Plugin::Check::CLI
     end
     http = Net::HTTP.new(settings['api']['host'], settings['api']['port'])
     req = net_http_req_class(method).new(path)
-    http.use_ssl = true if config[:ssl]
+    http.use_ssl = true if config[:protocol] == 'https'
     if settings['api']['user'] && settings['api']['password']
       req.basic_auth(settings['api']['user'], settings['api']['password'])
     end